tcfl: move to use the new power interface, keep backwards compat
With the introduction of the new power interface to ttbd, the client
needs to be able to use both the old and new one. Once all targets are
moved to the new one, we can deprecate the old. This, however, will
take time.

This moves all the command line handling for the power part to
target_ext_power and away from the common areas.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>